% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iita.R
\name{generate_quasiorders}
\alias{generate_quasiorders}
\title{Generate All Possible Quasi-Orders}
\usage{
generate_quasiorders(ni)
}
\arguments{
\item{ni}{Number of items}
}
\value{
A list of matrices, where each matrix represents a quasi-order as
  an adjacency matrix. Entry (i,j) = 1 means item i is a prerequisite for item j.
}
\description{
Generates all possible quasi-orders (prerequisite structures) for a given
number of items.
}
\details{
For computational efficiency, this function generates only
  the minimal set of prerequisite relations (the transitive reduction).
}
