% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iita.R
\name{print.iita}
\alias{print.iita}
\title{Print Method for IITA Objects}
\usage{
\method{print}{iita}(x, ...)
}
\arguments{
\item{x}{An object of class "iita"}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for IITA Objects
}
