% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iita.R
\name{transitive_closure}
\alias{transitive_closure}
\title{Compute Transitive Closure}
\usage{
transitive_closure(m)
}
\arguments{
\item{m}{A binary matrix representing a relation}
}
\value{
The transitive closure of m
}
\description{
Computes the transitive closure of a binary relation matrix.
}
\keyword{internal}
