% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iita.R
\name{print.iita_na}
\alias{print.iita_na}
\title{Print Method for IITA Objects}
\usage{
\method{print}{iita_na}(x, ...)
}
\arguments{
\item{x}{An object of class "iita_na"}

\item{...}{Additional arguments (not used)}
}
\description{
Prints a summary of the IITA analysis results. When multiple quasi-orders
are selected, it displays the most complex one (with the most relations)
as a helpful starting point for interpretation.
}
